[fact] move main.ts to make root folder cleaner
