+++ rollup.config.js
@@ -1,6 +1,7 @@
 import typescript from '@rollup/plugin-typescript';
 import {nodeResolve} from '@rollup/plugin-node-resolve';
 import commonjs from '@rollup/plugin-commonjs';
+import resolve from "@rollup/plugin-node-resolve";
 
 const isProd = (process.env.BUILD === 'production');
 
@@ -12,7 +13,7 @@ if you want to view the source visit the plugins github repository
 `;
 
 export default {
-  input: 'main.ts',
+  input: 'src/main.ts',
   output: {
     dir: '.',
     sourcemap: 'inline',
@@ -21,9 +22,12 @@ export default {
     exports: 'default',
     banner,
   },
-  external: ['obsidian'],
+  external: ["obsidian", "fs", "os", "path"],
   plugins: [
     typescript(),
+    resolve({
+      browser: true,
+    }),
     nodeResolve({browser: true}),
     commonjs(),
   ]
