+++ src/ui/BibleReferenceSettingTab.ts
@@ -1,6 +1,6 @@
 import { App, Notice, PluginSettingTab, Setting } from 'obsidian';
 import { APP_NAMING } from '../data/constants';
-import BibleReferencePlugin from '../../main';
+import BibleReferencePlugin from './../main';
 import { BibleVersionCollection, IBibleVersion } from '../data/BibleVersionCollection';
 
 export class BibleReferenceSettingTab extends PluginSettingTab {
